src/requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.3
beautifulsoup4==4.8.2
billiard==3.6.3.0
celery==4.4.2
Django==3.0.2
django-appconf==1.0.3
django-auth-ldap==2.1.0
django-compressor==2.4
django-cors-headers==3.2.1
django-filter==2.2.0
django-livereload-server==0.3.2
django-sass-processor==0.8
django-smtp-ssl==1.0
et-xmlfile==1.0.1
importlib-metadata==1.6.0
jdcal==1.4.1
kombu==4.6.8
libsass==0.19.4
mysql-client==0.0.1
mysqlclient==1.4.4
openpyxl==3.0.3
pdfkit==0.6.1
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-ldap==3.2.0
pytz==2019.3
rcssmin==1.0.6
redis==3.5.0
rjsmin==1.1.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
tornado==6.0.3
vine==1.3.0
XlsxWriter==3.0.1
xlwt==1.3.0
zipp==3.1.0
